refactor(indicator): move to script setup
