+++ src/config.json
@@ -286,6 +286,7 @@
           "name": "Indicator",
           "cName": "指示器",
           "desc": "显示一个任务或流程的进度，常用语开通流程。",
+          "setup": true,
           "author": "senyawang"
         },
         {
