+++ src/packages/__VUE/indicator/doc.en-US.md
@@ -38,6 +38,14 @@ app.use(Indicator)
 | align | Alignment, which takes effect only when `block` is `true`. optional value `left`, `right`, `center` | string | `left` |
 | fill-zero | Whether to add 0 before singular number | boolean | `true` |
 
+### Types version
+
+The component exports the following type definitions:
+
+```ts
+import type { IndicatorAlign, IndicatorProps, IndicatorInstance } from '@nutui/nutui'
+```
+
 ## Theming
 
 ### CSS Variables
